[2.x] 修复 server-only 与数据库 CLI 边界 (#376)
